Move the items into a panel
Separate control type, just lives inside the Listbox.
